backend/requirements.txt
torch==2.1.0
tensorflow==2.13.0
transformers==4.33.0
sentence-transformers==2.2.2
bertopic==0.15.0
nltk==3.8.1
spacy==3.7.2
PyPDF2==3.0.1
flask==2.0.1
flask-cors==3.0.10
google-generativeai==0.3.2pp